fix: permit only expected params
So that we don't forward other params to the Marvel API

In the future, the `Adapters::Marvel::RequestAdapter` class could potentially take care of that.

But for now let's keep it simple.

---

On another note, this controller is becoming too messy, but I need to focus on the frontend now, so will leave as it is.

At least there are tests o/